refactor: move Apple debug symbols under ios platform
