+++ src/client-types.ts
@@ -43,19 +43,13 @@ import type { ScreenshotRequestFlags } from './contracts/screenshot.ts';
 import type { PerfAction, PerfArea, PerfKind, PerfSubject } from './contracts/perf.ts';
 import type { DaemonBatchStep } from './core/batch.ts';
 import type { AlertAction, AlertInfo } from './alert-contract.ts';
-import type {
-  DebugSymbolsOptions,
-  DebugSymbolsResult,
-} from './commands/debugging/runtime/debug-symbols.ts';
+import type { DebugSymbolsOptions, DebugSymbolsResult } from './contracts/debug-symbols.ts';
 
 export type { FindLocator } from './utils/finders.ts';
 export type { CompanionTunnelScope, MetroBridgeScope } from './client-companion-tunnel-contract.ts';
 export type { AppsFilter } from './contracts/app-inventory.ts';
 export type { AlertAction, AlertInfo, AlertPlatform, AlertSource } from './alert-contract.ts';
-export type {
-  DebugSymbolsOptions,
-  DebugSymbolsResult,
-} from './commands/debugging/runtime/debug-symbols.ts';
+export type { DebugSymbolsOptions, DebugSymbolsResult } from './contracts/debug-symbols.ts';
 
 export type AgentDeviceDaemonTransport = (
   req: Omit<DaemonRequest, 'token'>,
