+++ src/platforms/ios/__tests__/debug-symbols.test.ts
@@ -3,9 +3,9 @@ import fs from 'node:fs/promises';
 import os from 'node:os';
 import path from 'node:path';
 import { test } from 'vitest';
-import { symbolicateCrashArtifact } from '../commands/debugging/runtime/debug-symbols.ts';
-import { AppError } from '../utils/errors.ts';
-import { withCommandExecutorOverride } from '../utils/exec.ts';
+import { symbolicateCrashArtifact } from '../debug-symbols.ts';
+import { AppError } from '../../../utils/errors.ts';
+import { withCommandExecutorOverride } from '../../../utils/exec.ts';
 
 const UUID = 'ABCDEFAB-CDEF-ABCD-EFAB-CDEFABCDEFAB';
 const NORMALIZED_UUID = 'ABCDEFABCDEFABCDEFABCDEFABCDEFAB';
