+++ src/platforms/ios/debug-symbols/crash-artifact.ts
