+++ src/platforms/ios/debug-symbols/report.ts
